Projectile Motion

+++ Projects/Projectile-Motion/ExperimentalData.py
@@ -1,5 +1,5 @@
 class ExperimentalData:
-    def __init__(self, pistol:str, caliber:str, ammunition:str, velocity:int, building:str, building_height:int, gravity:int):
+    def __init__(self, pistol:str, caliber:str, ammunition:str, velocity:int, building:str, building_height:int, gravity:float):
         self.pistol = pistol
         self.caliber = caliber
         self.ammunition = ammunition
